backend/requirements.txt
fastapi
uvicorn
pydantic-settings
supabase
openai
requests
python-dotenv